Cleaned up some multiplayer stuff
I’m getting a weird bug where if the boat in the bottom left corner
rotates about 45 degrees then all the letter keys and enter will stop
responding. Please tell me if you have any ideas about how to fix this.

+++ src/boatGamePackage/Sprite.java
@@ -40,7 +40,7 @@ public abstract class Sprite {
 	
 	// Draw the sprite to the screen
 	public void visualize() {
-		myDraw.picture(myX, myY, myFileName, myWidth, myHeight, myAngle);
+		myDraw.picture(myX, myY, myFileName, myWidth, myHeight, (int) myAngle);
 	}
 	
 	// Update the position of the sprite
